+++ src/canvas.class.ts
@@ -1,8 +1,8 @@
 //@ts-nocheck
 import { dragHandler, getActionFromCorner } from './controls/actions';
 import { Point } from './point.class';
-import { FabricObject } from './shapes/object.class';
-import { ModifierKey, Transform } from './typedefs';
+import { FabricObject } from './shapes/fabricObject.class';
+import { Transform } from './typedefs';
 import { saveObjectTransform } from './util/misc/objectTransforms';
 
 (function (global) {
@@ -347,23 +347,6 @@ import { saveObjectTransform } from './util/misc/objectTransforms';
        * @default
        */
       preserveObjectStacking: false,
-
-      /**
-       * Indicates the angle that an object will lock to while rotating.
-       * @type Number
-       * @since 1.6.7
-       * @default
-       */
-      snapAngle: 0,
-
-      /**
-       * Indicates the distance from the snapAngle the rotation will lock to the snapAngle.
-       * When `null`, the snapThreshold will default to the snapAngle.
-       * @type null|Number
-       * @since 1.6.7
-       * @default
-       */
-      snapThreshold: null,
 
       /**
        * Indicates if the right click on canvas can output the context menu or not